Major refactoring to separate header and delta application
This works by separating the header lookup, pack_object_header, from
other code which previously reconstituted the object at the same time.
If the object is deltified, the function will return the type. If the
object is deltified, it will identify the necessary values to patch the
object. These values can be used by pack_delta_content to reconstitute
the full object and size.

This process is further documented in lib/pack.c:249 and lib/pack.c:126